requirements.txt
openai==1.33.0
streamlit==1.36.0
python-dotenv==1.0.1